services/notification_api/requirements.txt
flask
flask-sqlalchemy
flask-mail
flask-restx
psycopg2-binary
requests
mailjet_rest